1.SONOFF Zigbee device OTA support (#1040)
* 1.SONOFF Zigbee device OTA support

* Clean up unnecessary fixes for unrelated OTA provider

* Split OTA provider unit tests into separate modules
