+++ Contributors.md
@@ -20,3 +20,4 @@
 - [https://github.com/Nemesis24] (https://github.com/Nemesis24)
 - [Hedda] (https://github.com/Hedda)
 - [Andreas Setterlind] (https://github.com/Gamester17)
+- [lisongjun] (https://github.com/lisongjun12)
