+++ tests/test_ota_provider.py
@@ -1,18 +1,10 @@
-import io
 import os.path
-import tarfile
 from unittest import mock
 import uuid
 
 import pytest
 
-from zigpy.config import (
-    CONF_OTA_DIR,
-    CONF_OTA_IKEA,
-    CONF_OTA_INOVELLI,
-    CONF_OTA_LEDVANCE,
-    CONF_OTA_SALUS,
-)
+from zigpy.config import CONF_OTA_DIR, CONF_OTA_IKEA
 import zigpy.ota
 import zigpy.ota.image
 import zigpy.ota.provider as ota_p
@@ -567,593 +559,3 @@ async def test_filestore_refresh_firmware_list_4(
     assert file_prov.update_expiration.call_count == 1
     assert len(file_prov._cache) == 1
     assert file_prov._cache[image_1.key].version == ver
-
-
-LEDVANCE_ID = 4489
-LEDVANCE_IMAGE_TYPE = 25
-
-
-@pytest.fixture
-def ledvance_prov():
-    p = ota_p.Ledvance()
-    p.enable()
-    return p
-
-
-@pytest.fixture
-def ledvance_image_with_version():
-    def img(version=100, image_type=LEDVANCE_IMAGE_TYPE):
-        img = zigpy.ota.provider.LedvanceImage(
-            LEDVANCE_ID, image_type, version, 180052, mock.sentinel.url
-        )
-        return img
-
-    return img
-
-
-@pytest.fixture
-def ledvance_image(ledvance_image_with_version):
-    return ledvance_image_with_version()
-
-
-@pytest.fixture
-def ledvance_key():
-    return zigpy.ota.image.ImageKey(LEDVANCE_ID, LEDVANCE_IMAGE_TYPE)
-
-
-async def test_ledvance_init_ota_dir(ledvance_prov):
-    ledvance_prov.enable = mock.MagicMock()
-    ledvance_prov.refresh_firmware_list = AsyncMock()
-
-    r = await ledvance_prov.initialize_provider({CONF_OTA_LEDVANCE: True})
-    assert r is None
-    assert ledvance_prov.enable.call_count == 1
-    assert ledvance_prov.refresh_firmware_list.call_count == 1
-
-
-async def test_ledvance_get_image_no_cache(ledvance_prov, ledvance_image):
-    ledvance_image.fetch_image = AsyncMock(return_value=mock.sentinel.image)
-    ledvance_prov._cache = mock.MagicMock()
-    ledvance_prov._cache.__getitem__.side_effect = KeyError()
-    ledvance_prov.refresh_firmware_list = AsyncMock()
-
-    # LEDVANCE manufacturer_id, but not in cache
-    assert ledvance_image.key not in ledvance_prov._cache
-    r = await ledvance_prov.get_image(ledvance_image.key)
-    assert r is None
-    assert ledvance_prov.refresh_firmware_list.call_count == 1
-    assert ledvance_prov._cache.__getitem__.call_count == 1
-    assert ledvance_image.fetch_image.call_count == 0
-
-
-async def test_ledvance_get_image(ledvance_prov, ledvance_key, ledvance_image):
-    ledvance_image.fetch_image = AsyncMock(return_value=mock.sentinel.image)
-    ledvance_prov._cache = mock.MagicMock()
-    ledvance_prov._cache.__getitem__.return_value = ledvance_image
-    ledvance_prov.refresh_firmware_list = AsyncMock()
-
-    r = await ledvance_prov.get_image(ledvance_key)
-    assert r is mock.sentinel.image
-    assert ledvance_prov._cache.__getitem__.call_count == 1
-    assert ledvance_prov._cache.__getitem__.call_args[0][0] == ledvance_image.key
-    assert ledvance_image.fetch_image.call_count == 1
-
-
-@patch("aiohttp.ClientSession.get")
-async def test_ledvance_refresh_list(
-    mock_get, ledvance_prov, ledvance_image_with_version
-):
-    ver1, img_type1 = (0x00102428, 25)
-    ver2, img_type2 = (0x00102428, 13)
-    img1 = ledvance_image_with_version(version=ver1, image_type=img_type1)
-    img2 = ledvance_image_with_version(version=ver2, image_type=img_type2)
-
-    sha_1 = "ffe0298312f63fa0be5e568886e419d714146652ff4747a8afed2de"
-    fn_1 = "A19 RGBW/00102428/A19_RGBW_IMG0019_00102428-encrypted"
-    sha_2 = "fa5ab550bde3e8c877cf40aa460fc9836405a7843df040e75bfdb2f"
-    fn_2 = "A19 TW 10 year/00102428/A19_TW_10_year_IMG000D_001024"
-    mock_get.return_value.__aenter__.return_value.json = AsyncMock(
-        side_effect=[
-            {
-                "firmwares": [
-                    {
-                        "blob": None,
-                        "identity": {
-                            "company": 4489,
-                            "product": 25,
-                            "version": {
-                                "major": 1,
-                                "minor": 2,
-                                "build": 428,
-                                "revision": 40,
-                            },
-                        },
-                        "releaseNotes": "",
-                        "shA256": sha_1,
-                        "name": "A19_RGBW_IMG0019_00102428-encrypted.ota",
-                        "productName": "A19 RGBW",
-                        "fullName": fn_1,
-                        "extension": ".ota",
-                        "released": "2019-02-28T16:36:28",
-                        "salesRegion": "us",
-                        "length": 180052,
-                    },
-                    {
-                        "blob": None,
-                        "identity": {
-                            "company": 4489,
-                            "product": 13,
-                            "version": {
-                                "major": 1,
-                                "minor": 2,
-                                "build": 428,
-                                "revision": 40,
-                            },
-                        },
-                        "releaseNotes": "",
-                        "shA256": sha_2,
-                        "name": "A19_TW_10_year_IMG000D_00102428-encrypted.ota",
-                        "productName": "A19 TW 10 year",
-                        "fullName": fn_2,
-                        "extension": ".ota",
-                        "released": "2019-02-28T16:42:50",
-                        "salesRegion": "us",
-                        "length": 170800,
-                    },
-                    # Old version but shows up after the new version in the OTA list
-                    {
-                        "blob": None,
-                        "identity": {
-                            "company": 4489,
-                            "product": 13,
-                            "version": {
-                                "major": 0,
-                                "minor": 2,
-                                "build": 428,
-                                "revision": 40,
-                            },
-                        },
-                        "releaseNotes": "",
-                        "shA256": sha_2,
-                        "name": "A19_TW_10_year_IMG000D_00102428-encrypted.ota",
-                        "productName": "A19 TW 10 year",
-                        "fullName": fn_2,
-                        "extension": ".ota",
-                        "released": "2015-02-28T16:42:50",
-                        "salesRegion": "us",
-                        "length": 170800,
-                    },
-                ]
-            }
-        ]
-    )
-    mock_get.return_value.__aenter__.return_value.status = 202
-    mock_get.return_value.__aenter__.return_value.reason = "OK"
-
-    await ledvance_prov.refresh_firmware_list()
-    assert mock_get.call_count == 1
-    assert len(ledvance_prov._cache) == 2
-    assert img1.key in ledvance_prov._cache
-    assert img2.key in ledvance_prov._cache
-    cached_1 = ledvance_prov._cache[img1.key]
-    assert cached_1.image_type == img1.image_type
-    base = "https://api.update.ledvance.com/v1/zigbee/firmwares/download"
-    assert cached_1.url == base + "?Company=4489&Product=25&Version=1.2.428.40"
-
-    cached_2 = ledvance_prov._cache[img2.key]
-    assert cached_2.image_type == img2.image_type
-    assert cached_2.url == base + "?Company=4489&Product=13&Version=1.2.428.40"
-
-    assert not ledvance_prov.expired
-
-
-@patch("aiohttp.ClientSession.get")
-async def test_ledvance_refresh_list_locked(
-    mock_get, ledvance_prov, ledvance_image_with_version
-):
-    await ledvance_prov._locks[ota_p.LOCK_REFRESH].acquire()
-
-    mock_get.return_value.__aenter__.return_value.json = AsyncMock(side_effect=[[]])
-
-    await ledvance_prov.refresh_firmware_list()
-    assert mock_get.call_count == 0
-
-
-@patch("aiohttp.ClientSession.get")
-async def test_ledvance_refresh_list_failed(mock_get, ledvance_prov):
-
-    mock_get.return_value.__aenter__.return_value.json = AsyncMock(side_effect=[[]])
-    mock_get.return_value.__aenter__.return_value.status = 434
-    mock_get.return_value.__aenter__.return_value.reason = "UNK"
-
-    with patch.object(ledvance_prov, "update_expiration") as update_exp:
-        await ledvance_prov.refresh_firmware_list()
-    assert mock_get.call_count == 1
-    assert update_exp.call_count == 0
-
-
-@patch("aiohttp.ClientSession.get")
-async def test_ledvance_fetch_image(mock_get, ledvance_image_with_version):
-    data = bytes.fromhex(
-        "1ef1ee0b0001380000008911012178563412020054657374204f544120496d61"
-        "676500000000000000000000000000000000000042000000"
-    )
-    sub_el = b"\x00\x00\x04\x00\x00\x00abcd"
-
-    img = ledvance_image_with_version(image_type=0x2101)
-    img.url = mock.sentinel.url
-
-    mock_get.return_value.__aenter__.return_value.read = AsyncMock(
-        side_effect=[data + sub_el]
-    )
-
-    r = await img.fetch_image()
-    assert isinstance(r, zigpy.ota.image.OTAImage)
-    assert mock_get.call_count == 1
-    assert mock_get.call_args[0][0] == mock.sentinel.url
-    assert r.serialize() == data + sub_el
-
-
-SALUS_ID = 4216
-SALUS_MODEL = "XY123"
-
-
-@pytest.fixture
-def salus_prov():
-    p = ota_p.Salus()
-    p.enable()
-    return p
-
-
-@pytest.fixture
-def salus_image_with_version():
-    def img(version=100, model=SALUS_MODEL):
-        img = zigpy.ota.provider.SalusImage(
-            SALUS_ID, model, version, 180052, mock.sentinel.url
-        )
-        return img
-
-    return img
-
-
-@pytest.fixture
-def salus_image(salus_image_with_version):
-    return salus_image_with_version()
-
-
-@pytest.fixture
-def salus_key():
-    return zigpy.ota.image.ImageKey(SALUS_ID, SALUS_MODEL)
-
-
-async def test_salus_init_ota_dir(salus_prov):
-    salus_prov.enable = mock.MagicMock()
-    salus_prov.refresh_firmware_list = AsyncMock()
-
-    r = await salus_prov.initialize_provider({CONF_OTA_SALUS: True})
-    assert r is None
-    assert salus_prov.enable.call_count == 1
-    assert salus_prov.refresh_firmware_list.call_count == 1
-
-
-async def test_salus_get_image_no_cache(salus_prov, salus_image):
-    salus_image.fetch_image = AsyncMock(return_value=mock.sentinel.image)
-    salus_prov._cache = mock.MagicMock()
-    salus_prov._cache.__getitem__.side_effect = KeyError()
-    salus_prov.refresh_firmware_list = AsyncMock()
-
-    # salus manufacturer_id, but not in cache
-    assert salus_image.key not in salus_prov._cache
-    r = await salus_prov.get_image(salus_image.key)
-    assert r is None
-    assert salus_prov.refresh_firmware_list.call_count == 1
-    assert salus_prov._cache.__getitem__.call_count == 1
-    assert salus_image.fetch_image.call_count == 0
-
-
-async def test_salus_get_image(salus_prov, salus_key, salus_image):
-    salus_image.fetch_image = AsyncMock(return_value=mock.sentinel.image)
-    salus_prov._cache = mock.MagicMock()
-    salus_prov._cache.__getitem__.return_value = salus_image
-    salus_prov.refresh_firmware_list = AsyncMock()
-
-    r = await salus_prov.get_image(salus_key)
-    assert r is mock.sentinel.image
-    assert salus_prov._cache.__getitem__.call_count == 1
-    assert salus_prov._cache.__getitem__.call_args[0][0] == salus_image.key
-    assert salus_image.fetch_image.call_count == 1
-
-
-@patch("aiohttp.ClientSession.get")
-async def test_salus_refresh_list(mock_get, salus_prov, salus_image_with_version):
-    img1 = salus_image_with_version(version="00000006", model="45856")
-    img2 = salus_image_with_version(version="00000006", model="45857")
-
-    mock_get.return_value.__aenter__.return_value.json = AsyncMock(
-        side_effect=[
-            {
-                "versions": [
-                    {
-                        "model": "45856",
-                        "version": "00000006",
-                        "url": "http://eu.salusconnect.io/download/firmware/a65779cd-13cd-41e5-a7e0-5346f24a0f62/45856_00000006.tar.gz",
-                    },
-                    {
-                        "model": "45857",
-                        "version": "00000006",
-                        "url": "http://eu.salusconnect.io/download/firmware/3319b501-98f3-4337-afbe-8d04bb9938bc/45857_00000006.tar.gz",
-                    },
-                ]
-            }
-        ]
-    )
-    mock_get.return_value.__aenter__.return_value.status = 202
-    mock_get.return_value.__aenter__.return_value.reason = "OK"
-
-    await salus_prov.refresh_firmware_list()
-    assert mock_get.call_count == 1
-    assert len(salus_prov._cache) == 2
-    assert img1.key in salus_prov._cache
-    assert img2.key in salus_prov._cache
-    cached_1 = salus_prov._cache[img1.key]
-    assert cached_1.model == img1.model
-    base = "http://eu.salusconnect.io/download/firmware/"
-    assert (
-        cached_1.url
-        == base + "a65779cd-13cd-41e5-a7e0-5346f24a0f62/45856_00000006.tar.gz"
-    )
-
-    cached_2 = salus_prov._cache[img2.key]
-    assert cached_2.model == img2.model
-    assert (
-        cached_2.url
-        == base + "3319b501-98f3-4337-afbe-8d04bb9938bc/45857_00000006.tar.gz"
-    )
-
-    assert not salus_prov.expired
-
-
-@patch("aiohttp.ClientSession.get")
-async def test_salus_refresh_list_locked(
-    mock_get, salus_prov, salus_image_with_version
-):
-    await salus_prov._locks[ota_p.LOCK_REFRESH].acquire()
-
-    mock_get.return_value.__aenter__.return_value.json = AsyncMock(side_effect=[[]])
-
-    await salus_prov.refresh_firmware_list()
-    assert mock_get.call_count == 0
-
-
-@patch("aiohttp.ClientSession.get")
-async def test_salus_refresh_list_failed(mock_get, salus_prov):
-
-    mock_get.return_value.__aenter__.return_value.json = AsyncMock(side_effect=[[]])
-    mock_get.return_value.__aenter__.return_value.status = 434
-    mock_get.return_value.__aenter__.return_value.reason = "UNK"
-
-    with patch.object(salus_prov, "update_expiration") as update_exp:
-        await salus_prov.refresh_firmware_list()
-    assert mock_get.call_count == 1
-    assert update_exp.call_count == 0
-
-
-@patch("aiohttp.ClientSession.get")
-async def test_salus_fetch_image(mock_get, salus_image_with_version):
-    data = bytes.fromhex(  # based on ikea sample but modded mfr code
-        "1ef1ee0b0001380000007810012178563412020054657374204f544120496d61"
-        "676500000000000000000000000000000000000042000000"
-    )
-
-    sub_el = b"\x00\x00\x04\x00\x00\x00abcd"
-    # construct tar.gz from header + sub_el
-    binstr = data + sub_el
-    fh = io.BytesIO()  # don't create a real file on disk, just in RAM.
-    with tarfile.open(fileobj=fh, mode="w:gz") as tar:
-        info = tarfile.TarInfo("salus_sample.ota")
-        info.size = len(binstr)
-        tar.addfile(info, io.BytesIO(binstr))
-
-    img = salus_image_with_version(model=SALUS_MODEL)
-    img.url = mock.sentinel.url
-
-    mock_get.return_value.__aenter__.return_value.read = AsyncMock(
-        side_effect=[fh.getvalue()]
-    )
-
-    r = await img.fetch_image()
-    assert isinstance(r, zigpy.ota.image.OTAImage)
-    assert mock_get.call_count == 1
-    assert mock_get.call_args[0][0] == mock.sentinel.url
-    assert r.serialize() == data + sub_el
-
-
-INOVELLI_ID = 4655
-INOVELLI_MODEL = "VZM31-SN"
-
-
-@pytest.fixture
-def inovelli_prov():
-    p = ota_p.Inovelli()
-    p.enable()
-    return p
-
-
-@pytest.fixture
-def inovelli_image_with_version():
-    def img(version=5, model=INOVELLI_MODEL):
-        img = zigpy.ota.provider.INOVELLIImage(
-            INOVELLI_ID, model, version, mock.sentinel.url
-        )
-        return img
-
-    return img
-
-
-@pytest.fixture
-def inovelli_image(inovelli_image_with_version):
-    return inovelli_image_with_version()
-
-
-@pytest.fixture
-def inovelli_key():
-    return zigpy.ota.image.ImageKey(INOVELLI_ID, INOVELLI_MODEL)
-
-
-async def test_inovelli_init_ota_dir(inovelli_prov):
-    inovelli_prov.enable = mock.MagicMock()
-    inovelli_prov.refresh_firmware_list = AsyncMock()
-
-    r = await inovelli_prov.initialize_provider({CONF_OTA_INOVELLI: True})
-    assert r is None
-    assert inovelli_prov.enable.call_count == 1
-    assert inovelli_prov.refresh_firmware_list.call_count == 1
-
-
-async def test_inovelli_get_image_no_cache(inovelli_prov, inovelli_image):
-    inovelli_image.fetch_image = AsyncMock(return_value=mock.sentinel.image)
-    inovelli_prov._cache = mock.MagicMock()
-    inovelli_prov._cache.__getitem__.side_effect = KeyError()
-    inovelli_prov.refresh_firmware_list = AsyncMock()
-
-    # inovelli manufacturer_id, but not in cache
-    assert inovelli_image.key not in inovelli_prov._cache
-    r = await inovelli_prov.get_image(inovelli_image.key)
-    assert r is None
-    assert inovelli_prov.refresh_firmware_list.call_count == 1
-    assert inovelli_prov._cache.__getitem__.call_count == 1
-    assert inovelli_image.fetch_image.call_count == 0
-
-
-async def test_inovelli_get_image(inovelli_prov, inovelli_key, inovelli_image):
-    inovelli_image.fetch_image = AsyncMock(return_value=mock.sentinel.image)
-    inovelli_prov._cache = mock.MagicMock()
-    inovelli_prov._cache.__getitem__.return_value = inovelli_image
-    inovelli_prov.refresh_firmware_list = AsyncMock()
-
-    r = await inovelli_prov.get_image(inovelli_key)
-    assert r is mock.sentinel.image
-    assert inovelli_prov._cache.__getitem__.call_count == 1
-    assert inovelli_prov._cache.__getitem__.call_args[0][0] == inovelli_image.key
-    assert inovelli_image.fetch_image.call_count == 1
-
-
-@patch("aiohttp.ClientSession.get")
-async def test_inovelli_refresh_list(
-    mock_get, inovelli_prov, inovelli_image_with_version
-):
-    img1 = inovelli_image_with_version(version="00000005", model="VMDS00I00")
-    img2 = inovelli_image_with_version(version="00000010", model="VZM31-SN")
-
-    base = "https://files.inovelli.com/firmware"
-
-    mock_get.return_value.__aenter__.return_value.json = AsyncMock(
-        side_effect=[
-            {
-                "VMDS00I00": [
-                    {
-                        "version": "00000005",
-                        "channel": "beta",
-                        "firmware": f"{base}/VZM31-SN/Beta/00000005/00000005.ota",
-                    }
-                ],
-                "VZM31-SN": [
-                    {
-                        "version": "00000005",
-                        "channel": "beta",
-                        "firmware": f"{base}/VZM31-SN/Beta/00000005/00000005.ota",
-                    },
-                    {
-                        "version": "00000006",
-                        "channel": "beta",
-                        "firmware": f"{base}/VZM31-SN/Beta/00000006/00000006.ota",
-                    },
-                    {
-                        "version": "00000010",
-                        "channel": "beta",
-                        "firmware": f"{base}/VZM31-SN/Beta/00000010/00000010.ota",
-                    },
-                    {
-                        "version": "00000007",
-                        "channel": "beta",
-                        "firmware": f"{base}/VZM31-SN/Beta/00000007/00000007.ota",
-                    },
-                    {
-                        "version": "00000008",
-                        "channel": "beta",
-                        "firmware": f"{base}/VZM31-SN/Beta/00000008/00000008.ota",
-                    },
-                    {
-                        "version": "00000009",
-                        "channel": "beta",
-                        "firmware": f"{base}/VZM31-SN/Beta/00000009/00000009.ota",
-                    },
-                ],
-            }
-        ]
-    )
-    mock_get.return_value.__aenter__.return_value.status = 202
-    mock_get.return_value.__aenter__.return_value.reason = "OK"
-
-    await inovelli_prov.refresh_firmware_list()
-    assert mock_get.call_count == 1
-    assert len(inovelli_prov._cache) == 2
-    assert img1.key in inovelli_prov._cache
-    assert img2.key in inovelli_prov._cache
-    cached_1 = inovelli_prov._cache[img1.key]
-    assert cached_1.model == img1.model
-    assert cached_1.url == f"{base}/VZM31-SN/Beta/00000005/00000005.ota"
-
-    cached_2 = inovelli_prov._cache[img2.key]
-    assert cached_2.model == img2.model
-    assert cached_2.url == f"{base}/VZM31-SN/Beta/00000010/00000010.ota"
-
-    assert not inovelli_prov.expired
-
-
-@patch("aiohttp.ClientSession.get")
-async def test_inovelli_refresh_list_locked(
-    mock_get, inovelli_prov, inovelli_image_with_version
-):
-    await inovelli_prov._locks[ota_p.LOCK_REFRESH].acquire()
-
-    mock_get.return_value.__aenter__.return_value.json = AsyncMock(side_effect=[[]])
-
-    await inovelli_prov.refresh_firmware_list()
-    assert mock_get.call_count == 0
-
-
-@patch("aiohttp.ClientSession.get")
-async def test_inovelli_refresh_list_failed(mock_get, inovelli_prov):
-
-    mock_get.return_value.__aenter__.return_value.json = AsyncMock(side_effect=[[]])
-    mock_get.return_value.__aenter__.return_value.status = 434
-    mock_get.return_value.__aenter__.return_value.reason = "UNK"
-
-    with patch.object(inovelli_prov, "update_expiration") as update_exp:
-        await inovelli_prov.refresh_firmware_list()
-    assert mock_get.call_count == 1
-    assert update_exp.call_count == 0
-
-
-@patch("aiohttp.ClientSession.get")
-async def test_inovelli_fetch_image(mock_get, inovelli_image_with_version):
-    header = bytes.fromhex(  # based on ikea sample but modded mfr code
-        "1ef1ee0b0001380000002f12012178563412020054657374204f544120496d61"
-        "676500000000000000000000000000000000000042000000"
-    )
-
-    sub_el = b"\x00\x00\x04\x00\x00\x00abcd"
-
-    img = inovelli_image_with_version(model=INOVELLI_MODEL)
-    img.url = mock.sentinel.url
-
-    mock_get.return_value.__aenter__.return_value.read = AsyncMock(
-        side_effect=[header + sub_el]
-    )
-
-    r = await img.fetch_image()
-    assert isinstance(r, zigpy.ota.image.OTAImage)
-    assert mock_get.call_count == 1
-    assert mock_get.call_args[0][0] == mock.sentinel.url
-    assert r.serialize() == header + sub_el
