+++ zigpy/config/defaults.py
@@ -18,6 +18,7 @@ CONF_OTA_INOVELLI_DEFAULT = False
 CONF_OTA_LEDVANCE_DEFAULT = False
 CONF_OTA_SALUS_DEFAULT = False
 CONF_OTA_OTAU_DIR_DEFAULT = None
+CONF_OTA_SONOFF_DEFAULT = False
 CONF_SOURCE_ROUTING_DEFAULT = False
 CONF_TOPO_SCAN_PERIOD_DEFAULT = 4 * 60  # 4 hours
 CONF_TOPO_SCAN_ENABLED_DEFAULT = True
